% Generated by roxygen2 (4.0.0): do not edit by hand
\name{distanceToPreviousVariant}
\alias{distanceToPreviousVariant}
\title{Find the distance between each variant and the previous one}
\usage{
distanceToPreviousVariant(vcf)
}
\arguments{
\item{vcf}{A \code{\link{VCF-class}}}
}
\value{
A vector of nrow(vcf)-1 giving the distance between
each variant and the previous one.
}
\description{
Given a \code{\link{VCF-class}} object, return a vector of distances
between neighboring variants.  This is useful for plotting defining
Kataegis phenomenon in a set of variants.  Distance for the first
variant on each chromosome is NA.
}
\examples{
require(VariantAnnotation)
fl <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
param <- ScanVcfParam(fixed="ALT", geno=c("GT", "GL"), info=c("LDAF"))
vcf = readVcf(fl,"hg19",param=param)
d = distanceToPreviousVariant(vcf)
summary(d)
}

