% Generated by roxygen2 (4.0.0): do not edit by hand
\name{isSNV}
\alias{isSNV}
\title{Determine if variants in a VCF object are SNPs}
\usage{
isSNV(vcf)
}
\arguments{
\item{VCF}{an object inheriting from the VCF class}
}
\value{
logical vector with the same length as \code{vcf}
}
\description{
Returns TRUE for variants that are SNVs and FALSE otherwise.
For variants with multiple ALT alleles, only the FIRST is used.
}
\examples{
library(VariantAnnotation)
fl <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
param <- ScanVcfParam(fixed="ALT", geno=c("GT", "GL"), info=c("LDAF"))
vcf = readVcf(fl,"hg19",param=param)
table(isSNV(vcf))
}
\seealso{
\code{\link{VCF-class}}
}
\keyword{character}

