% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{variantContext}
\alias{variantContext}
\alias{variantContext,VCF,FaFile-method}
\alias{variantContext,VRanges,FaFile-method}
\title{Gather variant context around single nucleotide variants}
\usage{
variantContext(VARIANTS, FASTA, ...)

\S4method{variantContext}{VCF,FaFile}(VARIANTS, FASTA, ...)

\S4method{variantContext}{VRanges,FaFile}(VARIANTS, FASTA, ...)
}
\arguments{
\item{variants}{A \code{\link{VCF-class}} object}

\item{fasta}{A \code{\link{FaFile}} object that describes the
fasta file location of the reference sequence}

\item{extendBy}{The number of bases on either side of the
altered base to capture.  The altered base will be placed in the
center of the context.  For example, the default value of 1
will result in contexts of length 3.  Specifying 2 will result
in contexts of length 5, etc.}
}
\value{
A \code{\link{VariantContext}} object.
}
\description{
The variant context is defined by gathering the nucleotides
upstream and downstream of a variant.  This if often the first
step in building a mutation spectrum analysis.
}
\section{Methods (by class)}{
\itemize{
\item \code{VARIANTS = VCF,FASTA = FaFile}: 

\item \code{VARIANTS = VRanges,FASTA = FaFile}: 
}}
\examples{
library(VariantAnnotation)
fl <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
param <- ScanVcfParam(info=c("DP"))
vcf = readVcf(fl,"hg19",param=param)
seqlevels(vcf) = paste0('chr',seqlevels(vcf))
variantContext(vcf,fasta=FaFile('/data/CCRBioinfo/public/GATK/bundle/2.3/hg19/ucsc.hg19.fasta'))
}

