% Generated by roxygen2 (4.0.0): do not edit by hand
\name{variantGCContent}
\alias{variantGCContent}
\title{Calculate GC content in window around a set of variants}
\usage{
variantGCContent(vcf, fasta, windowSize = 100)
}
\arguments{
\item{vcf}{A \code{\link{VCF-class}} object}

\item{fasta}{A \code{\link{FaFile}} object}

\item{windowSize}{An integer representing the window size around each variant
to calculate GC content}
}
\value{
A numeric vector giving the GC content for each variant window in the
input
}
\description{
Calculate GC content in window around a set of variants
}
\examples{
\dontrun{
require(VariantAnnotation)
require(AnnotationHub)
fl <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
param <- ScanVcfParam(fixed="ALT", geno=c("GT", "GL"), info=c("LDAF"))
vcf = readVcf(fl,"hg19",param=param)
ah = AnnotationHub()
hg19fa = ah$ensembl.release.72.fasta.homo_sapiens.dna.Homo_sapiens.GRCh37.72.dna.toplevel.fa.rz
varGC = variantGCContent(vcf,hg19fa)
plot(density(varGC))
}
}

